In the data2pg extension, rework the function that generates messages for DISCOVERY batches. There are now 4 types of messages: ERROR messages report about source data content that would produce an error if copied as is into the target database; CORRUPTION messages notice source data content that would be corrupted (truncated or precision loss), ADVICE messages suggest column type optimizations based on discovered data, and WARNING messages reports analysis limits like empty tables or nothing but NULL values. Rename the discovery_advice table into discovery_message. Drop the now useless discovery_column.dscv_max_size column. The missing analysis for float columns is left for another patch.

+++ test_pg/setup.sql
@@ -205,7 +205,7 @@ DROP TABLE IF EXISTS myTbl4 ;
 CREATE TABLE myTbl4 (
   col41       INT              NOT NULL,
   col42       TEXT             ,
-  col43       INT              ,
+  col43       NUMERIC          ,
   col44       DECIMAL(7)       ,
   col45       CHAR(10)         ,
   PRIMARY KEY (col41),
